Replace static context with Configuration class
