+++ spec/flyml_spec.rb
@@ -1,28 +1,28 @@
 require 'spec_helper'
 
 describe Flyml do
-  before do
-    Flyml.env = 'test'
-    Flyml.root = File.join(File.dirname(__FILE__), 'fixtures')
-  end
+  let(:env) { 'test' }
+  let(:root) { File.join(File.dirname(__FILE__), 'fixtures') }
+
+  subject(:config) { Flyml.config(root: root, env: env) }
 
   it 'accesses settings within a file' do
-    expect(Flyml[:simple]).to eq(foo: 'bar')
+    expect(config[:simple]).to eq(foo: 'bar')
   end
 
   it 'accesses nested settings' do
-    expect(Flyml[:nested][:this][:is]).to eq('nested')
+    expect(config[:nested][:this][:is]).to eq('nested')
   end
 
   it 'accesses empty files' do
-    expect(subject[:empty]).to eq({})
-    expect(subject[:empty][:inexisting]).to eq(nil)
+    expect(config[:empty]).to eq({})
+    expect(config[:empty][:inexisting]).to eq(nil)
   end
 
   it 'accesses files only the first time' do
-    expect(subject[:simple][:foo]).to eq('bar')
-    
+    expect(config[:simple][:foo]).to eq('bar')
+
     expect(YAML).not_to receive(:load_file)
-    expect(subject[:simple][:foo]).to eq('bar')
+    expect(config[:simple][:foo]).to eq('bar')
   end
 end
